chore: usePurchaseTest 훅의 반환값이 변경됨으로 ts -> tsx 파일로 전환
